requirements.txt
django>=3.1,<3.2
psycopg2-binary>=2.8
djangorestframework>=3.11,<3.12